Rename nextstrain profile to nextstrain-gisaid
For consistency with the nextstrain-open profile and the /ncov/gisaid/…
and /ncov/open/… dataset prefixes/URLs we'll be using.

Other profile names (rhino, scicore, etc) may want to change as well in
the future, to reflect that they're profiles for specific compute
environments instead of profiles for specific data sources and build
definitions.