Shift to per-build auspice configs
This mirrors the behavior from the "nextstrain-gisaid" profile
and is the simplest to reason with in the short term.
The config JSONs are missing the "title" key as this is
never used during the workflow as we provide a "--title" arg
to augur export.

There are very few differences between the configs here, and some
are identical. The changes introduced here are made in light of ongoing
discussions about implementing config overlays or producing the config
via a snakemake rule, both of which are out-of-scope for this PR.