Remove build sizes complexity
This reduces the build targets for the open builds to mirror
those in the nextstrain-gisaid profile, namely one-per-region.
This change is intended to facilitate the merge of this PR
without introducing too much additional complexity.

We will revisit the idea of programmatically generating multiple
builds, be they different focal groups, build sizes, or other
variables.