Update intermediate assets uploaded
The intermediate files from the "nextstrain-open" profile
will be available for public access, and represent a (soft, non-binding)
commitment from us to provided these on a semi-regular basis.

The three diagnostic files are now rarely used so we don't need
to upload them.

For each (subsampled) build we add a compressed version of the build
alignment, as this will allow an easier starting point for many people.
Finally, we upload the (per-build) auspice JSONs for completeness.