Remove ability to use diagnostic files as input
A previous commit modified what data we upload and removed 3 diagnostic
files (the output from `rule diagnostic`). This commit represents the
other side of this decision: it removes the ability for builds to
define those files as remote inputs (the "to-exclude" key of an input).

We also fix a potential bug in `_collect_exclusion_files()` as when
starting from a masked input we do not have the ability to generate
the exclusion file (as that needs access to the alignment).

Finally the `compress_exclusion_file` rule is removed as that was
only used for file upload.

Note that this commit does not affect the uploaded
`mutation-summary.tsv.xz` file.