Add Wuhan/Hu-1/2019 reference data
As of commit 7373a0a, we use Wuhan/Hu-1/2019 to root time trees, but we
did not update the `data/` references to include this sequence and
metadata. As a result, users following our ncov tutorial receive an
error at the `refine` step because the root sequence does not exist.
This commit adds the missing root metadata and sequence, but it keeps
the original Wuhan/WH01/2019 data in case other existing analyses depend
on it.