Include optional rule to prepend hCoV-19/ to strain names
GISAID requires that strain names be prepended by "hCoV-19/" for web applications that use GISAID data. This commit adds a script that takes in Auspice JSON and tip frequencies JSON and an arbitrary prefix and updates strains with this prefix. A rule is added to run this script and prefix with "hCoV-19/" when specified by the config parameter "include_hcov19_prefix". This config parameter defaults to False.

I chose to run rule regardless and toggle --prefix between "hCoV-19/" and "" in order to keep snakemake logic simpler. Before arriving at this implementation, I started with a similar logic to "skip_travel_history_adjustment", but this was compounding with "skip_travel_history_adjustment" and I thought better to just run script regardless and have the script either modify files or not depending on config setting.