upload-to-s3: Drop Content-Encoding from object metadata
This avoids the field being sent as an HTTP header, which causes most
HTTP clients to automatically decompress the content.  The automatic
decompression is useful if piping directly to another process, but
pretty confusing when downloading because the filename still retains the
.gz extension but the file isn't compressed!  This confuses both
browser-based downloads and the Broad's mirroring of S3 to GCP Storage.

Additionally, as Ivan and others have noted, ~no HTTP clients support
Content-Encoding: xz, even though we were setting it and the bulk of our
files are xz-compressed now.