docs: Emphasize distinction between sequences/aligned/masked/filtered even more
The current trio of properties is bolded as a whole, which is good, but
it still took me a few reads to notice they were slightly different, and
I'm _familiar_ with the workflow.  I thought the docs were wrong at
first!  They're correct, but emphasize the missing properties a bit more
to hopefully make the distinction stand out more.