Find clusters of closely related strains
Adds a script that finds polytomies in a phylogeny such that each polytomy has
at least N tips (N=3 by default) and all tips in the polytomy belong to the same
group based on strain-specific values from a predefined metadata column. By
default, all strains must belong to the same division to be considered part of
the same cluster. This default can be changed by the user as described in the
updated configuration guide. The resulting clusters appear as categorical
annotations in the final Auspice tree.