Annotate epiweek as color and filter option
Although we plan to eventually support custom start and end date values
through Auspice's filter interface, this commit some similar information
that is also relevant to epidemiologists in the form of CDC-style
epiweeks. We add a new script to calculate epiweeks using the Python
library of the same name and produce a node data JSON representation of
those values.

Note that while this commit updates the Conda environment file to
represent the new dependency on the "epiweeks" package, we would also
need to include this library in the Docker base image, if we end up
merging this code.

We would also need to update all auspice configs to include "epiweek" as
a color-by and filter. This commit updates only the default auspice
config.