Remove travel exposure from pipeline
In the initial stages of the pandemic we implemented an ad-hoc approach
to dealing with samples which were travel associated. At this time we
had few sequences, and were actively tracking the travel history for all
of them. This implementation required new metadata fields being
introduced, e.g. country_exposure, and the DTA was performed on this
data. We modified the tree such that terminal nodes actually became
2 nodes: an internal node with the country of exposure, and a single
child node with the country of sampling (ditto for other geographic
resolutions); a solution which was less than ideal for Auspice
(e.g. mutations went on one of those branches, so tips often appeared
to have no mutations!).

This commit changes the pipeline to not use this metadata information,
thus making all builds behave as if
`config["skip_travel_history_adjustment"]` was True. The Snakemake
file contains checks and warns users if their config file specifies
unused parameters. Requesting a trait reconstruction of an exposure
value is a fatal error to ensure users update their configs.

See GitHub issue #646 for more discussion on the rational for this
change.

This commit closes #646