Add preprocess GitHub actions
These actions are intended to be triggered via repository dispatch
events sent from the appropriate fetch-and-ingest actions in
nextstrain/ncov-ingest.

Testing reveals that the GISAID preprocessing takes around 6 hours
(and this cannot be sped up by throwing more cores at the problem)
so we detach immediately after submitting the AWS job so that we don't
run into the GitHub action's 6 hour job limit.

We may wish to have this preprocessing job (snakemake profile) trigger
an appropriate phylogenetic rebuild in the future, but this has not
been implemented here.

These actions can also be triggered through the GitHub UI, with the
user choosing whether to upload to a trial location or overwrite
the existing files.