Add rule to rename GISAID's PANGO lineage
GISAID stores PANGO lineages in different columns depending on the
download format of the metadata. In the "Input for Augur pipeline"
download from the search interface, the lineage is stored in a
"pangolin_lineage" column. We did not previously rename this column to
match the Nextstrain column of "pango_lineage", so builds that relied
heavily on search results from GISAID were missing the PANGO lineage
annotations.