Revert action through github interface;
creates a github action that allows
those with permissions to run GH actions
on this repository to rever either gisaid
or open builds (all regions) to a given
date-stamped set of builds on s3. Must be
run through the github interface similar
to how trial builds workflows work in this
repo.