docs: Replace orientation file tables with lists
Replaces tables of input, output, and other files with lists in the file
orientation docs. In addition to fixing broken links that were not being
properly rendered in Markdown, this change allows the user to see all of
the content without scrolling. In the previous implementation, much of
the table content was hidden on narrow displays and there was no visual
indication that users should scroll.

This commit also corrects the content of this page, fixing some typos
and adding links to external resources.

Fixes #734