GitHub Actions to rebuild datasets
This adds companion actions to those introduced in #730.

These allow us to manually rebuild the datasets (starting from
preprocessed files, i.e. filtered) using GitHub's UI as well as via
a repository dispatch endpoint. The latter is to be implemented in
a later commit.

Note that despite using bash's `set -x`, secrets are not printed
(echo'd) to the GitHub action logs, which is convenient as these are
publicly visible.

The trial-build is removed as that functionality can be accomplished
by these actions.