Upload rule fails if the underlying script fails
If the underlying script (upload-to-s3) failed, this exit code was
not being passed onto snakemake, instead we were getting the exit
code of `tee` (i.e. 0). This resulted in workflows appearing to succeed
but when looking deeper, the actual uploading failed.

This behaviour surprised me as the docs state that "shell commands in
Snakemake use the bash shell in strict mode by default." [1,2]

[1] https://snakemake.readthedocs.io/en/stable/project_info/faq.html?highlight=strict%20mode#id11
[2] https://snakemake.readthedocs.io/en/stable/snakefiles/rules.html