Allow preprocessing actions to trigger phylogenetic rebuilds
Existing preprocessing actions are being triggered by their
corresponding fetch-and-ingest action (in nextstrain/ncov-ingest), which
itself runs on a daily schedule.

The intention here is to allow successful preprocessing jobs to trigger
their own phylogenetic rebuilds when a new alignment is ready.

Note that the preprocessing jobs run on AWS; they are triggered by a
GitHub action but due to their runtime the action itself doesn't monitor
the job itself. This means the trigger for the subsequent rebuild cannot
come from the GitHub action.