Parameterize metadata chunk size
Add a user configuration parameter for the chunk size of metadata
consumed during the sanitize metadata step. Exposing this parameter
allows users to tune the value to meet their needs (for example, using a
higher value when more memory is available) and allows us to set a value
in our CI workflow that tests the ability of the sanitize metadata
script to accurate process multiple chunks of input from the smaller
test data.