Sanitize tar-based sequences with TextIOWrapper
Replaces a generator that decodes the tarfile's binary io.BufferedReader
stream with Python's TextIOWrapper [1]. This wrapper is Python's
high-level interface to consume buffered binary streams like the
io.BufferedReader. It provides a standard file-like interface including
`read` and `peek` methods, etc.

This change simplifies FASTA-file handling for tar inputs, but it also
fixes a bug that appears when using BioPython 1.79. Until recently, we
have pinned BioPython to 1.76 or earlier, meaning our latest BioPython
code dates back to December 2019. Since then, BioPython has rewritten
its file handling interface to include more sophisticated error handling
like checking whether files are opened in text mode or not [2]. This
error handling depends on the presence of a `read` method in the input
source. The original implementation of the tarfile reader in this ncov
script passed a generator to BioPython that then caused a `TypeError`
when looking for the generator's nonexistent `read` method.

Fixes #760

[1] https://docs.python.org/3/library/io.html#io.TextIOWrapper
[2] https://github.com/biopython/biopython/blob/d72560fd78c8953f28c9add4f7bbbf8bd34293d2/Bio/SeqIO/Interfaces.py#L49