Add GISAID's id field as default
Adds GISAID's "Accession ID" field as a default database id column for
the sanitize metadata step. This aligns the defaults in the script with
the values in `defaults/parameters.yaml` and fixes a bug caused by a
mismatch in this script and the data preparation guide where field
renaming used to happen before deduplication. Now that deduplication
occurs first, we need to support the original database id field names
instead of assuming the renamed fields are available.

Fixes #761