Allow metadata deduplication to be skipped
Changes the behavior of the sanitize metadata script such that missing
database id columns in the metadata produces a warning instead of an
error. The original behavior works when input data have database
accessions (e.g., GISAID or GenBank), but it failed for local input data
that has not been submitted to a database yet. This type of data
represents the use case for users who just sequenced genomes and want to
analyze them locally before submission to a database. The "multiple
inputs" guide in our ncov documentation represents this use case.

Fixes #766