Set correct mode when opening temp file
Support for (pandas to_csv) writing to binary files was introduced in
v.1.2.0. To support previous versions (which augur does), we open the
NamedTemporaryFile [1] specifically in `wt` mode. An explicit newline
was provided as per the pandas docs [2].

For posterity, the error message was
TypeError: a bytes-like object is required, not 'str'.

[1] https://docs.python.org/3/library/tempfile.html#tempfile.NamedTemporaryFile
[2] https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.to_csv.html#pandas.DataFrame.to_csv