fix: sorting prior to applying the crowing penalty inverted
for each focal set, we sort the sequences that picked it as closest neighbor by priority (after shuffling to break spurious patterns in the input data) and then spread these priorities using the crowding penalties such close matches to different parts of the focal set are included in the builds (i.e. have highest priorities). The problem, however, was that the initial sort was in reverse order such that the highest priority sequence ended up with the highest crowding penalty. This nullified the proximity priorities in large data sets.