Improve log output for sanitized metadata errors
Augur's `read_metadata` function can throw a generic Exception when the
input is missing id columns. This exception is not easily distinguished
from any other error, so we check the contents of the error message and
raise a more specific error for better handling of unexpected errors.
Specifically, when an unexpected error occurs, the user will get a full
Python traceback in the log file instead of a generic "ERROR" with the
exception message.