Notify slack using threads
**Rationale**
The slack messages were becoming problematic due to both the number and
the long-running nature of the job, which made it hard to link messages
to the pipeline which triggered them.

**Solution**
We now use one slack thread per pipeline invocation. Failure messages
are broadcast to the channel as well, so that they are harder to miss.
Message content has also been improved.

**Testing**
Tested running locally and posting to slack channel #scratch

**Implementation details**
Sending slack messages as a thread response requires knowing the
timestamp of the parent message. As snakemake (potentially) runs each
rule in a different process, we use pytools' PersistentDict to do this,
as recommended by the Snakemake docs [1]. To avoid a new dependency
(pytools) I chose to copy PersistentDict's source code [2] to this repo.
Note that contrary to [1], the PersistentDict persists upon future runs
of the pipeline, so we clear it when the pipeline ends. (As we typically
run this on fresh EC2 instances this shouldn't be an issue.)

[1] https://snakemake.readthedocs.io/en/stable/project_info/faq.html#i-want-to-pass-variables-between-rules-is-that-possible
[2] pytools 2021.2.9 at https://github.com/inducer/pytools/blob/a0d9fdc33f459802b54a79c00a617fbc410e9d4d/pytools/persistent_dict.py