export_for_nextstrain: Use config "build_date"
Use a config "build_date" if it's available, with the default being the
old behavior of running a function within the expand function to get the
current date. This allows users to define build date that is stable for
the entire run of the build. We add the "build_date" config to
the rebuild GitHub workflows to ensure that we don't run into missing
input file erros if a build runs over multiple days.