Fix input function exceptions
Fixes multiple input function exceptions that occur when running the
Nextstrain profiles with older versions of Snakemake (pre-5.20.0). Since
Nextstrain builds produce dated JSONs (e.g., with "YYYY-MM-DD.json"
suffixes), the change to allow numbers in build names makes these dated
JSONs ambiguous to create from Snakemake's perspective.

Despite the earlier inclusion of rule orders to prioritize the
"dated_json" rule over the now-ambiguous "finalize" rule, Snakemake
still needs to walk through the graph of possible ways to create the
dated JSON files before it accounts for the rule order. If any
exceptions occur during this graph walk, pre-5.20.0 Snakemake versions
would exit with an "InputFunctionException". As of 5.20.0, Snakemake
catches these exceptions internally and keeps searching the graph for
valid paths [1]. Because Snakemake initially interprets the value of the
"build_name" wildcard to be something like "europe_2021-11-24" instead
of "europe", all dictionary lookups in the workflow that expect "europe"
use the dated name instead and raise key errors. The solution in this
commit is to replace hardcoded key lookups with `get` method calls and a
default empty dictionary.

Although we should really update the Nextstrain Docker base image to use
the latest Snakemake (6.10.0 from Oct 2021) instead of 5.10.0 (from Jan
2020), this commit fixes the input function exceptions that were
breaking the CI build in the Docker image and likely the cause of the
error that @rneher mentioned in the original PR comments [2].

[1] https://github.com/snakemake/snakemake/pull/479
[2] https://github.com/nextstrain/ncov/pull/524#issuecomment-757539824