Fix multiple inputs with one or more remotes
Fixes a bug where a build with multiple inputs including at least one
remote file produced an invalid return argument for Snakemake in the
`combined_sequences_for_subsampling` rule. The issue was caused when the
remote URLs returned by `path_or_url` were lists instead of scalars. The
input function for the Snakemake rule performed a naive list
comprehension across all inputs that would produce a list with at least
one list for each remote file in it. This list of lists is not a valid
Snakemake input.

Fixes #838