Skip Nextclade annotations if they exist already
Checks for existing Nextclade QC columns in the given metadata and skips
merge of the metadata and given Nextclade QC file when these existing
annotations are complete. This check prevents overwriting of the
Nextstrain team's Nextclade annotations from ncov-ingest. The ingest
workflow always downloads and runs the latest Nextclade release, while
this workflow depends on the version in the Docker image or conda
environment that could be slightly outdated.

The check for existing annotations was originally complicated by the
renaming of columns by the sanitize metadata step where an original
column like `aaSubstitution` would be renamed to `aa_substitution`. We
could update the `column_map` in the join script to map `aaSubstitution`
to `aa_substitution`, but the production Nextstrain builds _skip_
sanitizing metadata, so they would retain the `aaSubstitution` column
and end up getting reannotated Nextclade QC metrics here. To simplify
this check a bit, this commit changes the sanitized column name for
amino acid substitutions to `aaSubstitution`.