Support older versions of Nextclade
Although we'd prefer to use the latest version of Nextclade, users may
have older versions installed, so we cannot assume their QC output will
match all of the columns found in the latest release. This commit uses
the intersection of QC columns and expected columns to support older
Nextclade versions.