Fix ambiguous date filtering for epiweeks
Replaces check for "X" ambiguous characters in dates with a more
conservative definition of a valid date based on a regular expression
applied to the string representation of the date column. This should
filter most incomplete date values including the original "2020-XX-XX"
type of date that we had previously filtered. These filters should avoid
errors in datetime conversion with pandas or epiweek conversion.

Fixes #845