Use a literal path for Nextclade data input
Replaces a reference to the _rule_ that produces Nextclade data needed
for alignment with the _literal path_ to the corresponding output. This
approach allows users to define custom rules that generate the same
output and override the workflow's default rules. This kind of
functionality is important, as it allows users to easily modify the
workflow DAG from their custom build configuration and avoid maintaining
a fork of the main workflow that can get outdated easily.

Related to #875