Update open profile to use default clade colorby
The Nextstrain GISAID profiles do not include an explicit "colorings" entry for "clade_membership". But by specifying "clade_membership" as default "color_by" this cause "Clade" to rise to be placed at the top of the Color By dropdown selection menu. 

In the current open builds, the default option of clade coloring is sorted to 5th entry from the top. 

This commit is intended to add consistency between GISAID and open builds and to place the default coloring at the top of the dropdown list.