Expose pages in main sidebar
This switches the sidebar to follow the style we use in the main
project (docs.nextstrain.org). I much prefer this for the "Tutorials"
and "Visualisation & Interpretation" sections. The "Reference
Material" section isn't perfect -- ideally the "[+]" icon would be
next to the caption (the capitilised title) alleviating the need
for a stub page /reference/index.html which is just the TOC; however
I don't believe this is possible in sphinx.