[docs] Split config params into sections
I think this general ordering makes sense (although I am open to
different section names!) and generally makes settings easier
to find. Searching for a value should work as before, but I imagine
a lot of people are skim-reading this page to get an overview of
the available configuration.

Outside of the TOC, the CSS we use makes it hard to distinguish
top-level config keys vs. parameters of those config objects, but that's
for another commit (or another PR!).

No actual details of config settings are changed here.