[config docs] reduce verbosity
As suggested in review by @victorlin.

On the surface we lose the distinction about what parameters were
required, however this was misleading as in practice all required
parameters are defined in our defaults, and Snakemake merges dicts
so that it is perfectly valid to only specify a subset of "required"
parameters.