Swap 10449A for 18163G
For whatever reason 10449A is fully congruent with Omicron clades in GISAID builds but not in GenBank builds. Swapping to 18163G makes things consistent across the board.

Although 18163G occurs on the branch leading to 21M, including it here for 21K and 21L helps to keep clades from becoming "inverted".