Standardize regex syntax
Use leading hyphen in regex character sets to follow the convention of
literal hyphens in either first or last position of a set and to support
easier appending of additional characters to the end of the set without
accidentally implementing a range with a trailing hyphen. Also, use a
compressed syntax for dates.