Template auspice config files for our core builds
The auspice config files were almost duplicates of each other, resulting
in 14 very similar JSONs across the open & genbank profiles. This made
it harder than it should be to change config values (e.g. see
https://github.com/nextstrain/ncov/pull/916) and also made it all-too-
easy for things to fall out-of-sync.

The adds a custom rule which is only run for nextstrain core builds.
The rule is intended to be human readable and makes it clear which
are the values which differ across open/genbank builds as well as across
the different builds within those profiles. The vast majority of the
config is unchanged.

There is one change in content made here: The orderings of {region, country,
division} colorings are now in that order for both profiles; previously
the open profile had them reversed.