Sanitize sequences prior to nextclade run
Sanitizes sequences, removing any duplicate sequences, prior to running
nextclade on subsampled data. This additional check should not add too
much to the runtime of the workflow, but it will allow the workflow to
continue running when input data contain duplicates and were not
sanitized by earlier steps in the workflow (e.g., full alignment or
multiple combined input sequences).