Remove conda reference from build_description rule
With "conda" in place I was getting the error:
```
RuleException in line 1368 of /Users/trvrb/Documents/src/ncov/workflow/snakemake_rules/main_workflow.smk:
Conda environments are only allowed with shell, script, notebook, or wrapper directives (not with run).
```
in the "build_description" rule.

I decided to just remove the reference to "conda" here as this is how the "rename_emerging_lineages" rule already works.