Revert name of docs environment file
The `conda env create` command described in the README for the docs
requires a file named `environment.yml`, by default. Unless RTD has its
own default environment file name that overrules this other default,
this change allows the original docs to work as expected.

Note, this commit switches `conda` to `mamba` in the README for faster
environment creation (and for consistency with install docs, etc. where
we already use `mamba`).