Add GISAID account registration to setup
Adds GISAID account registration as the first step of the
setup/installation for these tutorials. The "custom data" tutorial
asks the user to curate data from the GISAID search interface. This is a
bad time to learn you haven't signed up for an account yet and will
likely frustrate users. By describing our expectation of a GISAID
account at the beginning, we can set the user up for a more positive
experience.