Minor edits to the custom-data tutorial
Revises the introductory text to center the user and use a single line
per sentence (a convention that simplifies collaborative editing via
version control systems). Revises the second item in the prerequisites
section to parallel the usage in the first item and use a reference to
the internal section of the documentation previously referenced by
emphasis styling. Clarifies the meaning of the "all" subsampling scheme
and what Auspice parameters in the Auspice config JSON do.