Dynamically calculate collection date range for docs
Uses the built-in docutils "date" directive to get today's date, defines
a custom "timedelta" directive that returns today's date plus whatever
timedelta the user defines in the docs, and uses these two dates to
define the range of collection dates for the genomic surveillance
tutorial's GISAID curation section. Although these dates will be
statically compiled into the documentation each time it is
built (instead of dynamically calculated when the page loads), this
implementation at least allows us to keep the example date range as
up-to-date as the documentation itself. We can always rebuild the docs
to get updated dates.