Reduce total samples for genomic surveillance build
Reduces the total number of samples used as input for the genomic
surveillance build by an order of 100. The original build took ~1.5
hours to run with 3 CPUs, while the modified build took only 3 minutes.
Adds a step in the "next steps" page to direct users to increase the
sample size.