Revise mem_mb definitions to always result in an integer > 0
Previously they would result in mem_mb=0 when the input size in GiB was
less than 1, as the int() would truncate that to 0 which then was
propagated to the final result by multiplication via the constant
scaling factor.  This only impacted smaller builds, such as our CI, but
good to fix anyway.

Use ceil() instead of int() to ensure that we always at least result in
mem_mb=1 (although almost every process we run is going to have a larger
memory footprint than that in reality, so there's still a bit of
inaccuracy…).