Let the user know when a requested build does not exist
Fixes an issue where the user can request a Snakemake target from the
command line that does not match any of the builds they have defined in
their configuration and Snakemake raises an unhelpful `KeyError`. This
change looks for the build name in the config first and then raises an
exception with a more helpful error message, to let the user know what
they can do to fix the problem.