Fix color assignment for clades
The script assign-colors.py is supposed to only include colors for traits that are present for the relevant coloring in filtered / subsampled metadata. Ie not every "location" should be included in the produced colors.tsv file, only locations that are present in the filtered / subsampled metadata.

There was a bug that was causing this not to work for clades. This was because clades are labeled as "clade_membership" in the Auspice JSON and as "Nextstrain_clade" in the metadata. This necessitated special casing this look up in the assign-colors.py script. Without this fix, the script always includes all clades in its output making Omicron forever orange / red.