Add RBD level coloring (omicron only)
For background please see "Variant report 2022-10-17" section of
https://github.com/neherlab/SARS-CoV-2_variant-reports/blob/main/reports/variant_report_latest_draft.md
for details of lineage classification. Additional changes via
@corneliusroemer in https://github.com/nextstrain/ncov/pull/1018.

Note that the mutations are relative to BA.2 so we should really use
this for omicron-only builds, which are being developed elsewhere.
This is especially relevant for S:493 where we consider a reversion
to increase the level, meaning strains basal to BA.2 will be elevated
inappropriately.