Generate a representative 100k sample
A subsampled input dataset (metadata + sequences) is useful for local
testing / development purposes. It is not expensive to compute this
weekly, so a GitHub action is added to run each monday.
(No timestamps are used as it's not intended to be accessed
retrospectively.)

There was a previous `config["upload"]` parameter which was only used
by nextstrain team builds and was removed in 9abad2a473d6f6f2a454ad3fdd7599b0a4ed7cbb
(December 2021). This parameter remained in some build YAMLs and has
been removed here. The new usage of the parameter is now documented.

Helpful review comments and code from @corneliusroemer, @trvrb,
@victorlin and @tsibley (https://github.com/nextstrain/ncov/pull/1032).

Co-authored-by: Thomas Sibley <tsibley@fredhutch.org>