Skip RTD redirects syncing on fork repositories
RTD redirects are to be synced on every push to master in this
repository. This requires triggering the workflow on push to that
branch, and ideally a workflow-level condition to not run on forks.
There is no such workflow-level condition, so the next best thing is a
job-level condition. On forks, the workflow will still run on each push
to master but the job will be skipped instead of failing due to a
missing RTD token.