Include immune escape and ACE2 scores if available
Adds colorings to the default Auspice config JSON for immune escape and
ACE2 binding scores and updates the list of columns to merge into the
metadata from Nextclade's QC output. This change allows users to get
these scores in their builds by changing the `nextclade_dataset` option
in their build config to "sars-cov-2-21L", but it has no effect on the
behavior of the default dataset, "sars-cov-2".