Strip directories from tarball members
When extracting tarballs prior to sanitizing data, uses only the final
filename component of the path for each member in the tarball. This
strips any leading directories that might be in the path like a single
"/" in GISAID's tarballs for Augur that would cause files to be written
to the root level of the file system.