Pre-filter the inputs to the 21L-rooted builds with a custom rule
Filter rules in the config are applied _after_ subsampling, which poses
issues with reliably getting the desired number of sequences.  As @trvrb
wrote¹:

> In the workflow, the filter rule happens after the subsampling rules.
> This makes it so that if we ask for say 2560 in a sampling bucket, we'll
> lose >50% due to filtering out non-21L-descending clades.
>
> This could be solved by padding count targets to compensate, but this is
> hacky and the numbers will change as time goes on. Or the filter rule
> could be placed again before subsample, but we moved it afterwards for
> good reasons.

A few custom rules for the builds allows us to prefilter the full
dataset before subsampling.  Currently these rules are specific to our
GISAID data source, but they could be easily expanded to our Open data
sources too.

In the future we might also provide clade-partitioned subsets from
ncov-ingest², which we could use here instead with some adaptation of
the build config.

¹ <https://github.com/nextstrain/ncov/pull/1029#issue-1433903916>
² e.g. <https://github.com/nextstrain/ncov-ingest/pull/398>