Avoid labeling excluded clades on the 21L-rooted builds
This avoids incorrect labeling of the trees by `augur clades` if a
sequence happens to have mutations that define one of the excluded
clades.  This happens because there aren't earlier sequences for those
clades to be labeled due to their exclusion.  For example, we've seen
some spurious labelings of 19A because we exclude such sequences from
these builds.  Thanks to @emmahodcroft for debugging and explaining
this!¹

Since we define our clades hierarchically with inheritance, a small new
program is needed to fully expand those definitions so we can filter
them without worrying about filtering out their "parent" clades.

¹ <https://github.com/nextstrain/ncov/pull/1029#issuecomment-1499044262>