Start from actually-aligned sequences
Use aligned sequences as the aligned sequences input, rather than pass
off unaligned sequences as the aligned sequences input.

This should be inconsequential to workflow behaviour or results, but it
makes the config a bit more straightforward and less confusing.

In a quick dig thru history, it seems like ncov-ingest's
aligned.fasta.xz was not _quite_ available when we first switched our
profiles to use an "aligned" input instead of a "sequences" input.  The
original use of "aligned" with unaligned sequences was driven by run
time concerns and related to the move of the filtering step after the
subsampling step and move of the "preprocess" steps from this workflow
(ncov) to ncov-ingest.¹

Resolves <https://github.com/nextstrain/ncov/issues/1054>.

¹ <https://github.com/nextstrain/ncov/pull/814>
  <https://github.com/nextstrain/ncov/pull/823>