Rebuild the 21L-rooted builds on the "gisaid/rebuild" event
This is the event triggered by ncov-ingest when there's new GISAID data,
which also triggers the standard ancestral-rooted builds.

The specific "rebuild-gisaid-21L" event for this workflow seems
unnecessary since it would only be triggered manually, and in that case
I expect folks to use the GitHub Actions web UI instead.