Create locations DataFrame with a strain index
This is a follow-up fix to 2ec313589cf2cc19617e1fa5a5b19ed9c866b405.

Prior to that commit, DataFrames were not indexed by strain and used the
default counter index. The concat relied on both the locations and
metadata DataFrames having the same counter index. This was broken in
the linked commit when the metadata was switched to index on strain.

Solutions are to either (1) revert to using counter index for both, (2)
use strain index for both, or (3) ignore the index when concatenating.

A strain index has benefits and is used elsewhere in the script, so this
commit implements solution (2).