Update arguments to augur clades
This updates the workflow to use the new clades interface from augur
v22 (see nextstrain/augur#728). In the process we can remove two
rules from the workflow. The minimum augur version is bumped to 22.0.1,
as that includes a couple of important bug-fixes.

If this workflow is run with augur prior to v22, the emerging_lineages
rule  will error due to unknown arguments.

The script add_branch_labels.py is no longer used and thus removed here
(as recommended in code review: https://github.com/nextstrain/ncov/pull/1000#discussion_r1195787805)
Note that it contained unused functionality to export spike mutations;
if we reinstate this in the future we should update the output format
to produce a node-data JSON with a custom branch label to avoid modifying
the auspice JSON after export.