Restore branch label ordering
This is a minor aesthetic change which restores the order of the
branch labels in Auspice so that clade appears before emerging_lingeage.
The change in ordering was introduced by the previous commit as
we moved from a custom script to using `augur clades` interface from
Augur v22.0.0

The ordering is not specified and cannot be configured, but in practice
it is the order of the keys in the dictionary (and which is stable in
CPython 3.6, Python 3.7, and most (all?) JS engines).