Bump versions for Nextstrain tools in Conda env
Bumps versions for Augur, Nextalign, and Nextclade along with Python. We
can't use Python 3.9 or higher with this environment file, since the
pinned versions of pangolin require older versions of dependencies that
aren't compatible with Python 3.9 or higher. In future updates, we could
either bump pangolin and pangolearn and/or define pangolin dependencies
in their own environment file and use that file in the workflow's few
pangolin rules. This latter option is probably better, since it
simplifies the default environment (speeding up its installation,
reducing its size) and still supports pangolin in the workflow with
Snakemake's `--use-conda` interface.