Check for empty user-defined subsampling scheme
Adds a check for an empty user-defined subsampling scheme and provides a
helpful error message when this happens to replace the unhelpful
`AttributeError` message users would get before when the workflow tried to
access a nonexistent method of a `NoneType`.

This error can happen when a user defines a custom subsampling scheme
where the contents are not properly indented below the `subsampling`
key. The resulting config is still a valid YAML file, so the emptiness
of the subsampling scheme needs to be checked specifically. We could
probably handle this better and more generically with a more stringent
YAML schema.