Specify Read the Docs build OS
Read the Docs will soon require the specification of a build OS in our
configuration file [1] and have already started temporary enforcement of
this requirement as a warning. This commit follows the example in the
RTD docs for conda environments [2], adding a `build` block using the
latest supported Ubuntu and mambaforge.

[1] https://blog.readthedocs.com/use-build-os-config/
[2] https://docs.readthedocs.io/en/stable/config-file/v2.html#conda