assign_rbd_levels: Update `basal_clade_label`
The clade labels were updated in https://github.com/nextstrain/ncov/pull/1065.
Update the clade label so that the `assign_rbd_levels` script can find
the correct basal clade.

I had considered pulling this value out as a parameter in the config YAML,
but the original commit message¹ implies that this should _not_ be
configurable.

¹ https://github.com/nextstrain/ncov/commit/fb5f44e1b935f80940c510d196a8a70e7057705b